fix: エディター警告一掃（Pylance・Ruff統合）
FastAPI http_server.py のインポート構造を簡潔化し、型チェック統合を改善。

- FastAPI http_server.py: 直接インポート化でHTTPException型を確保
- financial_independence.py: list[int]→list[float]型アノテーション修正
- logging_config.py: structlog optional dependency に type: ignore 追加
- chart_generator.py: matplotlib optional dependency に type: ignore 追加
- database/__init__.py: __all__ 動的定義エラーを type: ignore で抑制

結果:
- ✅ Ruff: All checks passed!
- ✅ Pylance warnings: 12 → 0 (100% 削減)
- 🟡 Pylance errors: 227 → 137 (40% 削減)

Fixes #NFR-017

+++ backend/src/household_mcp/web/routes/financial_independence.py
@@ -107,7 +107,7 @@ async def get_financial_independence_projections(
         # TODO: 実際のデータソースから取得
         current_assets = 5000000
         annual_expense = 1000000
-        asset_history = [5000000 + (i * 50000) for i in range(period_months)]
+        asset_history = [float(5000000 + (i * 50000)) for i in range(period_months)]
 
         scenarios = analyzer.calculate_scenarios(
             current_assets=current_assets,
@@ -164,9 +164,9 @@ async def get_expense_breakdown(
     try:
         # TODO: 実際のデータソースから取得
         category_history = {
-            "食費": [50000 + (i * 100) for i in range(period_months)],
-            "交通費": [5000 if i % 2 == 0 else 0 for i in range(period_months)],
-            "医療費": [200000 if i == 5 else 0 for i in range(period_months)],
+            "食費": [float(50000 + (i * 100)) for i in range(period_months)],
+            "交通費": [float(5000 if i % 2 == 0 else 0) for i in range(period_months)],
+            "医療費": [float(200000 if i == 5 else 0) for i in range(period_months)],
         }
 
         classification_results = analyzer.classify_expenses(
